+++ internal/models/memory_item.go
@@ -17,6 +17,6 @@ type MemoryItem struct {
 	ArchivedAt *time.Time `json:"archivedAt,omitempty" yaml:"archivedAt,omitempty"`
 
 	EmbeddingProviderModelName *string    `json:"embeddingProviderModelName,omitempty" yaml:"embeddingProviderModelName,omitempty"`
-	Embedding                  *[]float32 `json:"embedding,omitempty" yaml:"embedding,omitempty"`
+	Embedding                  *[]float64 `json:"embedding,omitempty" yaml:"embedding,omitempty"`
 	EmbeddedAt                 *time.Time `json:"embeddedAt,omitempty" yaml:"embeddedAt,omitempty"`
 }
